Prioritize syntax for updates builder and make return type more specific
